+++ crates/lns-cli/src/cli.rs
@@ -248,71 +248,6 @@ fn parse_injection(s: &str) -> Result<lns_policy::providers::InjectionDef, Strin
 }
 
 #[derive(clap::Args)]
-pub struct PolicyArgs {
-    #[command(subcommand)]
-    pub command: PolicyCommand,
-}
-
-#[derive(Subcommand)]
-pub enum PolicyCommand {
-    #[command(about = "Add an allow rule for a destination pattern.")]
-    Allow(PolicyRuleArgs),
-    #[command(about = "Add a deny rule for a destination pattern.")]
-    Deny(PolicyRuleArgs),
-    #[command(about = "List the rules in the policy file.")]
-    List(PolicyScopeArgs),
-    #[command(about = "Remove the rule matching a destination pattern.")]
-    Remove(PolicyRemoveArgs),
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, ValueEnum)]
-#[value(rename_all = "lower")]
-pub enum TransportArg {
-    Direct,
-    Upstream,
-}
-
-#[derive(clap::Args)]
-pub struct PolicyRuleArgs {
-    #[arg(help = "Destination pattern: host, wildcard (*.example.com), CIDR, or host:port.")]
-    pub pattern: String,
-    #[arg(long, help = "Human-readable note stored alongside the rule.")]
-    pub description: Option<String>,
-    #[arg(
-        long,
-        value_enum,
-        default_value_t = TransportArg::Direct,
-        help = "Transport for an allowed connection; ignored for deny rules."
-    )]
-    pub transport: TransportArg,
-    #[arg(
-        long,
-        help = "Policy file path; defaults to `lns-policy.yaml` in the current directory."
-    )]
-    pub policy: Option<PathBuf>,
-}
-
-#[derive(clap::Args)]
-pub struct PolicyScopeArgs {
-    #[arg(
-        long,
-        help = "Policy file path; defaults to `lns-policy.yaml` in the current directory."
-    )]
-    pub policy: Option<PathBuf>,
-}
-
-#[derive(clap::Args)]
-pub struct PolicyRemoveArgs {
-    #[arg(help = "Destination pattern of the rule to remove.")]
-    pub pattern: String,
-    #[arg(
-        long,
-        help = "Policy file path; defaults to `lns-policy.yaml` in the current directory."
-    )]
-    pub policy: Option<PathBuf>,
-}
-
-#[derive(clap::Args)]
 pub struct RunArgs {
     #[arg(
         help = "OCI image reference (e.g. alpine:3.20); omit for imageless mode, which requires a command after `--`."
